Implement btp section
- modify nid, messageCount parameter/return type from int32 to int64
- reorganize network type module interface
- add BitSetFilter
- modify btp section and btp digest interface
- implement btp section from digest
- reorganize section builder

+++ block/blockv2.go
@@ -272,7 +272,7 @@ func (b *blockV2) BTPSection() module.BTPSection {
 	panic("implement me")
 }
 
-func (b *blockV2) BTPBlockFor(nid int32) (module.BTPBlock, error) {
+func (b *blockV2) BTPBlockFor(nid int64) (module.BTPBlock, error) {
 	//TODO implement me
 	panic("implement me")
 }
