+++ common/legacy/block_v0.go
@@ -195,7 +195,7 @@ func (b *blockV0) BTPSection() module.BTPSection {
 	panic("implement me")
 }
 
-func (b *blockV0) BTPBlockFor(nid int32) (module.BTPBlock, error) {
+func (b *blockV0) BTPBlockFor(nid int64) (module.BTPBlock, error) {
 	//TODO implement me
 	panic("implement me")
 }
